tiny_stories/requirements.txt
torch
transformers
networkx
datasets
kronfluence
numpy
einops